+++ R/CohortConstruction.R
@@ -557,6 +557,15 @@ instantiateCohortSet <- function(connectionDetails = NULL,
           results_database_schema.cohort_inclusion_stats = "#cohort_inc_stats",
           results_database_schema.cohort_summary_stats = "#cohort_summary_stats"
         )
+        # added for compatibility for 2.8.1
+        # https://github.com/OHDSI/CohortDiagnostics/issues/387
+        # this table was introduced in v2.8.1, and does not exist in prior version of webapi
+        if (stringr::str_detect(string = sql,
+                                pattern = 'cohort_censor_stats')) {
+          sql <- SqlRender::render(sql = sql,
+                                   results_database_schema.cohort_censor_stats = "#cohort_censor_stats"
+                                   )
+        }
       } else {
         sql <- SqlRender::render(
           sql,
@@ -729,7 +738,8 @@ saveAndDropTempInclusionStatsTables <- function(connection,
     if (any(
       stringr::str_detect(string = sql, pattern = "_inclusion_result"),
       stringr::str_detect(string = sql, pattern = "_inclusion_stats"),
-      stringr::str_detect(string = sql, pattern = "_summary_stats")
+      stringr::str_detect(string = sql, pattern = "_summary_stats"),
+      stringr::str_detect(string = sql, pattern = "_censor_stats")
     )) {
       if (isFALSE(generateInclusionStats)) {
         warning(